Move trait to persist batches to the storage workspace
