+++ node/libs/storage/src/lib.rs
@@ -1,5 +1,6 @@
 //! Abstraction for persistent data storage.
 //! It provides schema-aware type-safe database access.
+mod batch_store;
 mod block_store;
 pub mod proto;
 mod replica_store;
@@ -8,6 +9,7 @@ pub mod testonly;
 mod tests;
 
 pub use crate::{
+    batch_store::PersistentBatchStore,
     block_store::{BlockStore, BlockStoreRunner, BlockStoreState, PersistentBlockStore},
     replica_store::{Proposal, ReplicaState, ReplicaStore},
 };
